requirements.txt
dagshub==0.2.18.post1
deepchecks==0.17.3
hydra-core==1.3.2
hyperopt==0.2.7
joblib==1.2.0
mlflow==2.4.1
numpy==1.24.3
omegaconf==2.3.0
pandas==2.0.2
pandera==0.15.1
patsy==0.5.3
pytest_steps==1.8.0
scikit_learn==1.2.2
xgboost==1.7.5
